backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.5
pydantic-settings==2.9.1
aiofiles==24.1.0
python-multipart==0.0.20
arq==0.28.0
chardet==7.4.3
FlagEmbedding==1.4.0
google-genai>=1.0.0
google-generativeai==0.8.6
groq==1.2.0
langdetect==1.0.9
langfuse==4.5.0
llama-index==0.14.21
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-magic==0.4.27
qdrant-client==1.17.1
ragas==0.4.3
redis==5.2.1
surya-ocr==0.17.1
tiktoken==0.12.0
transformers>=4.46.0,<5.0.0